VoiceManagementSystem/requirement.txt
Django==2.0
django-bootstrap3==9.1.0
djangorestframework==3.7.7
Faker==0.8.8
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
text-unidecode==1.1
